refactor: update server to point to api/
